Check in support for the client side cache
Actually, it is unclear whether we even need a client side cache since it looks
like switching to HTTP GET implements server side caching anyway. While testing
this, I discovered that turning off the client side cache, or commenting it out
from the Application class still does not pull javascript files when they are
changed.

At this point, it is in fact very hard to invalidate the cache - the primary
way that I found is to add a version to the javascript SCRIPT tags that forces
a refresh of the file whose version changed.

A client side cache will also avoid validation requests, which we don't
actually track with our current end to end setup, so let's just include it anyway.

Note also that the background fetch part of the code works even less because it
doesn't actually parse the HTML to fetch the javascript, and attempts to load
the information in a webview doesn't work.