Demo of code to use plugins and to load parts of the interface from the server
a) use plugins (the statusbar plugin)
b) load parts of the interface from the server

We should really come up with a better sync mechanism for (b), but in that
absence of that, the HTTP cache will probably be good enough, specially given
that we probably won't change too much once the study has been deployed.

Need to test carefully though.

Next up:
- try out the database plugin
- do a poor man's template by replacing <remote_url> in the file that we are
  loading with the configured remote URL