Copy over some of the cordova-specific changes to the AppDelegate
The cordova app delegate has some code required for smooth cordova operation.
In particular, it enables the HTTP cache, and sends URL handling notifications
to all plugins.

We do not include the UI window creation since that does not work - it results
in a black screen. That is reflected in the documentation.

We need to make a similar change to dispatch remote notifications to the
plugins as well.